requirements.txt
Django==4.1.7
django-extensions==3.2.1
flake8==6.0.0
flake8-builtins==2.1.0
flake8-import-order==0.18.2
flake8-polyfill==1.0.2
ipython==8.10.0
pep8-naming==0.13.3
python-dotenv~=0.21.1
Faker~=17.0.0
webargs==8.2.0
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
decorator==5.1.1
Django==4.1.7
django-extensions==3.2.1
django-phone-field==1.8.1
executing==1.2.0
Faker==17.0.0
flake8==6.0.0
flake8-builtins==2.1.0
flake8-import-order==0.18.2
flake8-polyfill==1.0.2
ipython==8.10.0
jedi==0.18.2
marshmallow==3.19.0
matplotlib-inline==0.1.6
mccabe==0.7.0
packaging==23.0
parso==0.8.3
pep8==1.7.1
pep8-naming==0.13.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.36
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.14.0
python-dateutil==2.8.2
python-dotenv==0.21.1
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
traitlets==5.9.0
wcwidth==0.2.6
webargs==8.2.0

iso639~=0.1.4
django-filter~=23.1
django-debug-toolbar~=4.0.0